+++ RouteService/App/Dto/DriverDto.cs
@@ -1,0 +1,7 @@
+public class DriverDto
+{
+    public string Id { get; set; }
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
+    public string HourlyRate { get; set; }
+}
