+++ RouteService/App/Dto/VehicleDto.cs
@@ -1,0 +1,9 @@
+
+public class VehicleDto
+{
+    public string Id { get; set; }
+    public string PlateNumber { get; set; }
+    public double FuelEfficiency { get; set; }
+    public double AverageFuelEfficiency { get; set; }
+    public bool IsUnderMaintenance { get; set; }
+}
