+++ RouteService/App/Interface/ITripGrpcService.cs
@@ -1,0 +1,11 @@
+
+using RouteService.Infrastructure.Proto.Client;
+
+namespace RouteService.App.Interface;
+public class ITripGrpcService : IDisposable
+{
+    public void Dispose()   
+    {
+        throw new NotImplementedException();
+    }
+}
