--- RouteService/App/Interface/IVehicleClient.cs
@@ -1,9 +1,0 @@
-﻿using RouteService.Infrastructure.Proto.Client;
-
-namespace RouteService.App.Interface
-{
-    public interface IVehicleClient
-    {
-        Task<VehicleDto> GetVehicleAsync(string vehicleId);
-    }
-}
