+++ RouteService/App/Interface/IVehicleGrpcClient.cs
@@ -1,0 +1,7 @@
+﻿namespace RouteService.App.Interface
+{
+    public interface IVehicleClient : IDisposable
+    {
+        Task<VehicleDto> GetVehicleAsync(string vehicleId);
+    }
+}
