+++ RouteService/Infrastructure/Proto/Client/DriverGrpcClient.cs
@@ -8,64 +8,60 @@ using MyApp.Grpc;
 using RouteService.App.Interface;
 
 namespace RouteService.Infrastructure.Proto.Client
+{
+    public class DriverGrpcClient : IDriverGrpcClient, IDisposable
     {
-        public class DriverGrpcClient : IDriverGrpcClient, IDisposable
+        private readonly ILogger<DriverGrpcClient> _logger;
+        private readonly IConfiguration _configuration;
+        private readonly GrpcChannel _channel;
+
+        public DriverGrpcClient(ILogger<DriverGrpcClient> logger, IConfiguration configuration)
         {
-            private readonly ILogger<DriverGrpcClient> _logger;
-            private readonly IConfiguration _configuration;
-            private readonly GrpcChannel _channel;
+            _logger = logger;
+            _configuration = configuration;
 
-            public DriverGrpcClient(ILogger<DriverGrpcClient> logger, IConfiguration configuration)
+            var driverServiceUrl = _configuration.GetValue<string>("Services:DriverService");
+            if (string.IsNullOrEmpty(driverServiceUrl))
             {
-                _logger = logger;
-                _configuration = configuration;
-
-                var driverServiceUrl = _configuration.GetValue<string>("Services:DriverService");
-                if (string.IsNullOrEmpty(driverServiceUrl))
-                {
-                    throw new ArgumentException("La URL del servicio Driver no está configurada.");
-                }
-
-                _channel = GrpcChannel.ForAddress(driverServiceUrl);
+                throw new ArgumentException("La URL del servicio Driver no está configurada.");
             }
 
-            public async Task<DriverDto> GetDriverAsync(string driverId)
+            _channel = GrpcChannel.ForAddress(driverServiceUrl);
+        }
+
+        public async Task<DriverDto> GetDriverAsync(string driverId)
+        {
+            try
             {
-                try
+                _logger.LogInformation($"Solicitando datos del conductor con ID: {driverId}");
+
+                var client = new DriverService.DriverServiceClient(_channel);
+                var response = await client.GetDriverAsync(new GetDriverRequest { Id = driverId });
+
+                var driverDto = new DriverDto
                 {
-                    _logger.LogInformation($"Solicitando datos del conductor con ID: {driverId}");
+                    Id = response.Id,
+                    FirstName = response.FirstName,
+                    LastName = response.LastName,
+                    HourlyRate = response.HourlyRate
+                };
 
-                    var client = new DriverService.DriverServiceClient(_channel);
-                    var response = await client.GetDriverAsync(new GetDriverRequest { Id = driverId });
-
-                    var driverDto = new DriverDto
-                    {
-                        Id = response.Id,
-                        FirstName = response.FirstName,
-                        LastName = response.LastName,
-                    };
-
-                    _logger.LogInformation($"Datos del conductor obtenidos correctamente para el ID: {driverId}");
-                    return driverDto;
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, $"Error al obtener datos del conductor con ID: {driverId}");
-                    throw;
-                }
+                _logger.LogInformation($"Datos del conductor obtenidos correctamente para el ID: {driverId}");
+                return driverDto;
             }
-
-            public void Dispose()
+            catch (Exception ex)
             {
-                _channel?.Dispose();
+                _logger.LogError(ex, $"Error al obtener datos del conductor con ID: {driverId}");
+                throw;
             }
         }
 
-        public class DriverDto
+        public void Dispose()
         {
-            public string Id { get; set; }
-            public string FirstName { get; set; }
-            public string LastName { get; set; }
+            _channel?.Dispose();
         }
     }
+
+
+}
